refactor: refactor setInterval apiThrottle
